It's a full game (almost)
The day cycle loops now and all planned features are present. (Excluding sound, animation, bug fixing, etc.). There is an end of day screen that has a xp counter that has an animation and shows level ups and gives the palyer the ability to choose a skill on every other level up. Then the day resets and the day loops.